backend/requirements.txt
anyio==3.6.2
APScheduler==3.10.0
asyncpg==0.27.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
Faker==17.3.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
Pyrogram==2.0.99
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uvicorn==0.20.0
